# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://127.0.0.1:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/servicos/{id}":{"put":{"tags":["servico-controller"],"summary":"Alterar um serviço","operationId":"alterar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServicoIncluirRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServicoResponse"}}}}}}},"/pecas/{id}":{"get":{"tags":["peca-controller"],"summary":"Buscar uma peça por ID","operationId":"buscarPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PecaResponse"}}}}}},"put":{"tags":["peca-controller"],"summary":"Atualizar uma peça","operationId":"alterar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PecaAtualizarRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PecaResponse"}}}}}}},"/veiculos":{"post":{"tags":["veiculo-controller"],"summary":"Cadastrar um veículo","operationId":"cadastrar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VeiculoIncluirRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VeiculoResponse"}}}}}}},"/usuarios":{"post":{"tags":["usuario-controller"],"summary":"Cadastrar um usuário","operationId":"cadastrar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioIncluirRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsuarioResponse"}}}}}}},"/servicos":{"post":{"tags":["servico-controller"],"summary":"Cadastrar um serviço","operationId":"cadastrar_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServicoIncluirRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServicoResponse"}}}}}}},"/prestacao-servico":{"post":{"tags":["prestacao-servico-controller"],"summary":"Cadastrar Prestação de Serviço","operationId":"cadastrar_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrestacaoServicoIncluirRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PrestacaoServicoResponse"}}}}}}},"/pecas":{"post":{"tags":["peca-controller"],"summary":"Cadastrar uma peça","operationId":"cadastrar_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PecaIncluirRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PecaResponse"}}}}}}},"/ordem-servicos":{"post":{"tags":["ordem-servico-controller"],"summary":"Cadastrar uma Ordem de Serviço","operationId":"cadastrar_5","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrdemServicoIncluirRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemServicoResponse"}}}}}}},"/clientes":{"post":{"tags":["cliente-controller"],"summary":"Cadastrar um cliente","operationId":"cadastrar_6","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClienteIncluirRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClienteResponse"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/alocacao-pecas":{"post":{"tags":["alocacao-peca-controller"],"summary":"Cadastro de uma Alocação de Peças","operationId":"cadastrar_7","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlocacaoPecaIncluirRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AlocacaoPecaResponse"}}}}}}},"/servicos/{id}/inativar":{"patch":{"tags":["servico-controller"],"summary":"Inativar um serviço","operationId":"inativar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServicoResponse"}}}}}}},"/servicos/{id}/ativar":{"patch":{"tags":["servico-controller"],"summary":"Ativar um serviço","operationId":"ativar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServicoResponse"}}}}}}},"/prestacao-servico/{id}/inativar":{"patch":{"tags":["prestacao-servico-controller"],"summary":"Inativar Prestação de Serviço por Id","operationId":"inativar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PrestacaoServicoResponse"}}}}}}},"/prestacao-servico/{id}/finalizar":{"patch":{"tags":["prestacao-servico-controller"],"operationId":"finalizar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PrestacaoServicoResponse"}}}}}}},"/ordem-servicos/{id}/voltar-diagnostico":{"patch":{"tags":["ordem-servico-controller"],"summary":"Voltar a ordem de serviço para diagnóstico","operationId":"voltarDiagnostico","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemServicoResponse"}}}}}}},"/ordem-servicos/{id}/pagar":{"patch":{"tags":["ordem-servico-controller"],"summary":"Pagar e entregar uma Ordem de Serviço","operationId":"pagarEEntregar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemServicoResponse"}}}}}}},"/ordem-servicos/{id}/iniciar-execucao":{"patch":{"tags":["ordem-servico-controller"],"summary":"Iniciar execução da ordem de serviço","operationId":"iniciarExecucao","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemServicoResponse"}}}}}}},"/ordem-servicos/{id}/enviar-para-aguardar-aprovacao":{"patch":{"tags":["ordem-servico-controller"],"summary":"Iniciar aguardando aprovação da ordem de serviço","operationId":"enviarParaAguardarAprovacao","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemServicoResponse"}}}}}}},"/ordem-servicos/{id}/cancelar":{"patch":{"tags":["ordem-servico-controller"],"summary":"Cancelar ordem de serviço","operationId":"cancelar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemServicoResponse"}}}}}}},"/clientes/{id}":{"patch":{"tags":["cliente-controller"],"summary":"Alterar um cliente","operationId":"alterar_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClienteAlterarRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClienteResponse"}}}}}}},"/alocacao-pecas/{id}/inativar":{"patch":{"tags":["alocacao-peca-controller"],"summary":"Inativação de uma Alocação de Peças","operationId":"inativar_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AlocacaoPecaResponse"}}}}}}},"/veiculos/{id}":{"get":{"tags":["veiculo-controller"],"summary":"Buscar um veículo por ID","operationId":"buscarPorId_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VeiculoResponse"}}}}}}},"/veiculos/por-placa/{placa}":{"get":{"tags":["veiculo-controller"],"summary":"Buscar um veículo por placa","operationId":"buscarPorPlaca","parameters":[{"name":"placa","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VeiculoResponse"}}}}}}},"/veiculos/por-cliente/{clienteId}":{"get":{"tags":["veiculo-controller"],"summary":"Buscar veículos por cliente ID","operationId":"buscarPorClienteId","parameters":[{"name":"clienteId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VeiculoResponse"}}}}}}}},"/prestacao-servico/por-ordem-servico/{ordemServicoId}":{"get":{"tags":["prestacao-servico-controller"],"summary":"Buscar todos Prestação de Serviços por Ordem Serviço ID","operationId":"buscarTodosPorOrdemServicoId","parameters":[{"name":"ordemServicoId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PrestacaoServicoResponse"}}}}}}}},"/ordem-servicos/{id}":{"get":{"tags":["ordem-servico-controller"],"summary":"Buscar uma Ordem de Serviço por ID","operationId":"buscarPorId_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdemServicoResponse"}}}}}}},"/ordem-servicos/{id}/tempo-medio":{"get":{"tags":["ordem-servico-controller"],"summary":"Buscar tempo médio dos serviços da Ordem de serviços por ID","operationId":"buscarTempoMedioDosServicos","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TempoMedioOSResponse"}}}}}}},"/ordem-servicos/veiculo/placa/{placa}":{"get":{"tags":["ordem-servico-controller"],"summary":"Buscar Ordem de serviços por placa do veículo","operationId":"buscarTodosPorPlaca","parameters":[{"name":"placa","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrdemServicoResponse"}}}}}}}},"/clientes/por-usuario/{usuarioId}":{"get":{"tags":["cliente-controller"],"summary":"Busca todos os clientes cadastrados por um usuário","operationId":"buscarTodosPorUsuarioId","parameters":[{"name":"usuarioId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClienteResponse"}}}}}}}},"/clientes/por-cpf/{cpf}":{"get":{"tags":["cliente-controller"],"summary":"Buscar cliente por CPF","operationId":"buscarPorCpf","parameters":[{"name":"cpf","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClienteResponse"}}}}}}},"/clientes/por-cnpj/{cnpj}":{"get":{"tags":["cliente-controller"],"summary":"Buscar cliente por CNPJ","operationId":"buscarPorCnpj","parameters":[{"name":"cnpj","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClienteResponse"}}}}}}}},"components":{"schemas":{"ServicoIncluirRequest":{"required":["nome"],"type":"object","properties":{"nome":{"type":"string"},"descricao":{"type":"string"}}},"ServicoResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"statusRecursoEnum":{"type":"string","enum":["Ativo","Inativo"]},"nome":{"type":"string"},"descricao":{"type":"string"}}},"PecaAtualizarRequest":{"required":["marca","nome","quantidadeEstoque","valorUnitario"],"type":"object","properties":{"nome":{"type":"string"},"marca":{"type":"string"},"valorUnitario":{"type":"number"},"quantidadeEstoque":{"type":"integer","format":"int32"}}},"PecaResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["Ativo","Inativo"]},"nome":{"type":"string"},"marca":{"type":"string"},"valorUnitario":{"type":"number"},"quantidadeEstoque":{"type":"integer","format":"int32"}}},"VeiculoIncluirRequest":{"required":["ano","clienteId","marca","modelo","placa","quantidadeEixos"],"type":"object","properties":{"placa":{"type":"string"},"marca":{"type":"string"},"modelo":{"type":"string"},"ano":{"type":"string"},"quantidadeEixos":{"type":"integer","format":"int32"},"clienteId":{"type":"string","format":"uuid"}}},"VeiculoResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["Ativo","Inativo"]},"placa":{"type":"string"},"marca":{"type":"string"},"modelo":{"type":"string"},"ano":{"type":"string"},"quantidadeEixos":{"type":"integer","format":"int32"},"clienteId":{"type":"string","format":"uuid"}}},"UsuarioIncluirRequest":{"required":["cargoEnum","email","nome","senha"],"type":"object","properties":{"nome":{"type":"string"},"email":{"type":"string"},"senha":{"type":"string"},"cargoEnum":{"type":"string","enum":["ATENDENTE","MECANICO","ALMOXARIFE","ATENDENTE, MECANICO, ALMOXARIFE"]}}},"UsuarioResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"statusRecursoEnum":{"type":"string","enum":["Ativo","Inativo"]},"nome":{"type":"string"},"email":{"type":"string"},"cargoEnum":{"type":"string","enum":["ATENDENTE","MECANICO","ALMOXARIFE"]}}},"PrestacaoServicoIncluirRequest":{"required":["ordemServicoId","precoMaoDeObra","servicoId"],"type":"object","properties":{"precoMaoDeObra":{"type":"number"},"ordemServicoId":{"type":"string","format":"uuid"},"servicoId":{"type":"string","format":"uuid"}}},"PrestacaoServicoResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["Ativo","Inativo"]},"precoMaoDeObra":{"type":"number"},"subtotal":{"type":"number"},"dataInicio":{"type":"string","format":"date-time"},"dataFim":{"type":"string","format":"date-time"},"ordemServicoId":{"type":"string","format":"uuid"},"servicoId":{"type":"string","format":"uuid"}}},"PecaIncluirRequest":{"required":["marca","nome","quantidadeEstoque","valorUnitario"],"type":"object","properties":{"nome":{"type":"string"},"marca":{"type":"string"},"valorUnitario":{"type":"number"},"quantidadeEstoque":{"type":"integer","format":"int32"}}},"OrdemServicoIncluirRequest":{"required":["veiculoId"],"type":"object","properties":{"observacao":{"type":"string"},"veiculoId":{"type":"string","format":"uuid"}}},"OrdemServicoResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["Ativo","Inativo"]},"situacao":{"type":"string","enum":["Recebida","Em diagnóstico","Aguardando aprovação","Em execução","Finalizado","Entregue"]},"dataRecebida":{"type":"string","format":"date-time"},"dataDiagnostico":{"type":"string","format":"date-time"},"dataAguardandoAprovacao":{"type":"string","format":"date-time"},"dataExecucao":{"type":"string","format":"date-time"},"dataFinalizada":{"type":"string","format":"date-time"},"dataEntregue":{"type":"string","format":"date-time"},"pago":{"type":"boolean"},"valorTotal":{"type":"number"},"observacao":{"type":"string"},"veiculoId":{"type":"string","format":"uuid"},"usuarioId":{"type":"string","format":"uuid"}}},"ClienteIncluirRequest":{"required":["email","nome"],"type":"object","properties":{"nome":{"maxLength":255,"minLength":3,"type":"string"},"cpf":{"type":"string"},"cnpj":{"type":"string"},"email":{"type":"string"},"telefone":{"type":"string"}}},"ClienteResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["Ativo","Inativo"]},"nome":{"type":"string"},"cpf":{"type":"string"},"cnpj":{"type":"string"},"email":{"type":"string"},"telefone":{"type":"string"},"usuarioId":{"type":"string","format":"uuid"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"senha":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"jwt":{"type":"string"}}},"AlocacaoPecaIncluirRequest":{"required":["pecaId","prestacaoServicoId","quantidadeNecessaria"],"type":"object","properties":{"quantidadeNecessaria":{"type":"integer","format":"int32"},"prestacaoServicoId":{"type":"string","format":"uuid"},"pecaId":{"type":"string","format":"uuid"}}},"AlocacaoPecaResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["Ativo","Inativo"]},"quantidadeNecessaria":{"type":"integer","format":"int32"},"prestacaoServicoId":{"type":"string","format":"uuid"},"pecaId":{"type":"string","format":"uuid"}}},"ClienteAlterarRequest":{"type":"object","properties":{"nome":{"type":"string"},"telefone":{"type":"string"}}},"TempoMedioOSResponse":{"type":"object","properties":{"servicos":{"type":"array","items":{"$ref":"#/components/schemas/TempoMedioServicoItemResponse"}},"tempoTotal":{"type":"string"}}},"TempoMedioServicoItemResponse":{"type":"object","properties":{"nomeServico":{"type":"string"},"tempoMedio":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}